game: Add base code, Initialise setup and main game loop, Define classes for Game and Dino
